Maximum Sum_Interval_Algorithm_Project
The maximum sum subarray problem involves finding the contiguous subarray
within a one-dimensional array of numbers that has the largest sum. Develop
an algorithm using either Divide and Conquer or Dynamic Programming,
excluding Brute Force, to efficiently solve this problem. The algorithm should
adhere to the following properties:
1. If the array contains all non-positive numbers, the solution is the number
in the array with the smallest magnitude.
2. If the array contains all non-negative numbers, the problem is trivial, and
the maximum sum is the sum of all the elements in the list.
3. An empty set is not a valid input.
4. There can be multiple different subarrays that achieve the same
maximum sum for the problem.
Example: For the array [-5, 6, 7, 1, 4, -8, 16], the maximum sum is 26, achieved
by adding 6 + 7 + 1 + 4 + -8 + 16